fix: move error defs
