+++ src/valust/src/error/mod.rs
@@ -1,0 +1,68 @@
+//! Human-readable error produced by `valust`.
+
+pub mod transform;
+pub mod validate;
+
+use std::fmt::{Debug, Display};
+
+use transform::TransformError;
+use validate::ValidateError;
+
+/// Display-able error trait.
+///
+/// You don't need to manually implement this trait as this trait
+/// has been implemented for all `Debug + Display` types.
+pub trait ErrorShow: Debug + Display {}
+
+impl<T: Debug + Display> ErrorShow for T {}
+
+/// Any validation error.
+#[derive(Debug, Default)]
+pub struct ValidationError {
+    /// Error produced by validators.
+    pub validates: Vec<ValidateError>,
+    /// Error produced by transformers.
+    pub transforms: Vec<TransformError>,
+}
+
+impl ValidationError {
+    /// Create an empty error set.
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// Check if the error set contains any error instance.
+    pub fn check(self) -> Result<(), ValidationError> {
+        if self.validates.is_empty() && self.transforms.is_empty() {
+            Ok(())
+        } else {
+            Err(self)
+        }
+    }
+
+    /// Push a validator error to the set.
+    pub fn push_validate_error(&mut self, err: ValidateError) {
+        self.validates.push(err);
+    }
+
+    /// Push a transformer error to the set.
+    pub fn push_transform_error(&mut self, err: TransformError) {
+        self.transforms.push(err);
+    }
+
+    /// Extend the set.
+    ///
+    /// This will modify original set's `path` field.
+    pub fn extend_error(&mut self, parent: &str, rhs: Self) {
+        self.validates
+            .extend(rhs.validates.into_iter().map(|mut x| {
+                x.path = format!("{}.{}", parent, x.path);
+                x
+            }));
+        self.transforms
+            .extend(rhs.transforms.into_iter().map(|mut x| {
+                x.path = format!("{}.{}", parent, x.path);
+                x
+            }));
+    }
+}
